requirements.txt
numpy==1.26.4
tensorflow==2.10.1
dagshub
dvc
ensure
fastapi
gdown
GitPython
joblib
python_box
PyYAML
scikit_learn
seaborn
tf2onnx
uvicorn
mlflow
onnx
python-multipart
matplotlib
GPUtil